requirements.txt
discord.py==2.3.2
openai==1.33.0
python-dotenv==1.0.1
